use `MAX_ARGS` for `Node::Function` array length

+++ src/evaluate/function/mod.rs
@@ -14,7 +14,7 @@ where
     T: Function<T>,
 {
     const NAMESPACE: &'static str;
-    const MAX_ARGS: u8;
+    const MAX_ARGS: usize;
     fn from_string(namespaces: &mut Iter<&str>, identifier: &str) -> Result<T, SimdevalError>;
     fn call(&self, node: &[Value]) -> Result<Value, SimdevalError>;
     fn is_const(&self) -> bool {
